Update: Refactored test suites for Help, Wallet Selection Page, Invalid Bitcoin Address, and Asset Invoice to align with recent updates

+++ src/data/repository/setting_card_repository.py
@@ -72,7 +72,7 @@ class SettingCardRepository:
         try:
             local_store.set_value('defaultFeeRate', rate)
             # Verify the setting was applied
-            if local_store.get_value('defaultFeeRate', value_type=float):
+            if local_store.get_value('defaultFeeRate', value_type=int):
                 return IsDefaultFeeRateSet(is_enabled=True)
             return IsDefaultFeeRateSet(is_enabled=False)
         except Exception as exe:
